Moved test for promise equivalence.
