+++ package.json
@@ -4,7 +4,7 @@
   "description": "A NodeJS package that provides implementations of some data structures",
   "main": "index.js",
   "scripts": {
-    "test": "mocha ./test/structures --recursive"
+    "test": "mocha ./test --recursive"
   },
   "repository": {
     "type": "git",
